+++ src/core/server/commands/getVehicle.ts
@@ -2,6 +2,7 @@ import * as alt from 'alt-server';
 import { addCommand } from '../systems/chat';
 import { getDescription } from '../../shared/commands/commandList';
 import { getVectorInFrontOfPlayer } from '../utility/vector';
+import { CommandsLocale } from '../../shared/locale/commands';
 
 addCommand('getvehicle', handleCommand);
 
@@ -12,7 +13,7 @@ function handleCommand(player: alt.Player, name: string): void {
     }
 
     if (player.data.isDead) {
-        player.send(`Unable to spawn vehicles while dead.`);
+        player.send(CommandsLocale.CANNOT_WHILE_DEAD);
         return;
     }
 
@@ -21,6 +22,6 @@ function handleCommand(player: alt.Player, name: string): void {
     try {
         const veh = new alt.Vehicle(name, forwardPosition.x, forwardPosition.y, forwardPosition.z, 0, 0, 0);
     } catch (err) {
-        player.send(`Vehicle model is not valid`);
+        player.send(CommandsLocale.VEHICLE_MODEL_NOT_VALID);
     }
 }