+++ src/core/shared/enums/animation.ts
@@ -1,0 +1,8 @@
+export enum AnimationFlags {
+    NORMAL = 0,
+    REPEAT = 1,
+    STOP_LAST_FRAME = 2,
+    UPPERBODY_ONLY = 16,
+    ENABLE_PLAYER_CONTROL = 32,
+    CANCELABLE = 120
+}