+++ src/core/shared/locale/commands.ts
@@ -1,0 +1,12 @@
+export enum CommandsLocale {
+    //
+    ARMOUR_SET_TO = 'Your armour was set to: ',
+    HEALTH_SET_TO = 'Your health was set to: ',
+    CANNOT_FIND_PLAYER = 'Could not find that player.',
+    CANNOT_WHILE_DEAD = 'Cannot perform this command while you are dead.',
+    ON = 'ON',
+    OFF = 'OFF',
+    PLAYER_IS_TOO_FAR = 'That player is too far away.',
+    PLAYER_IS_NOT_DEAD = 'That player is not dead.',
+    VEHICLE_MODEL_NOT_VALID = 'Vehicle model is not vehicle.'
+}